refactor: game catalog hooks to hook
