+++ src/scenes/Catalog/hooks/game-catalog.hook.ts
